std::move collections when emplacing them into events
