+++ RecoEgamma/EgammaElectronProducers/plugins/LowPtGsfElectronSeedValueMapsProducer.cc
@@ -8,7 +8,29 @@
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
 #include "FWCore/Utilities/interface/InputTag.h"
-#include "RecoEgamma/EgammaElectronProducers/plugins/LowPtGsfElectronSeedValueMapsProducer.h"
+#include "DataFormats/Common/interface/ValueMap.h"
+#include "DataFormats/GsfTrackReco/interface/GsfTrackFwd.h"
+#include "DataFormats/ParticleFlowReco/interface/PreIdFwd.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+
+#include <vector>
+#include <string>
+
+class LowPtGsfElectronSeedValueMapsProducer : public edm::stream::EDProducer<> {
+public:
+  explicit LowPtGsfElectronSeedValueMapsProducer(const edm::ParameterSet&);
+
+  void produce(edm::Event&, const edm::EventSetup&) override;
+
+  static void fillDescriptions(edm::ConfigurationDescriptions&);
+
+private:
+  const edm::EDGetTokenT<reco::GsfTrackCollection> gsfTracks_;
+  const edm::EDGetTokenT<edm::ValueMap<reco::PreIdRef> > preIdsValueMap_;
+  const std::vector<std::string> names_;
+};
 
 ////////////////////////////////////////////////////////////////////////////////
 //
@@ -20,10 +42,6 @@ LowPtGsfElectronSeedValueMapsProducer::LowPtGsfElectronSeedValueMapsProducer(con
     produces<edm::ValueMap<float> >(name);
   }
 }
-
-////////////////////////////////////////////////////////////////////////////////
-//
-LowPtGsfElectronSeedValueMapsProducer::~LowPtGsfElectronSeedValueMapsProducer() {}
 
 ////////////////////////////////////////////////////////////////////////////////
 //
